exif_burn/requirements.txt
pillow==11.0.0
PyExifTool==0.5.6